requirements.txt
build
twine
pyjnius
setuptools
